Move hook & nav back to view recipe screen
